develop part of section 6

+++ config/config.go
@@ -63,7 +63,7 @@ type ScreenshotAPI struct {
 	Path                     string `json:"path"`
 	UpdateFrozenDurationPath string `json:"update_frozen_duration_path"`
 	FrozenDuration           int    `json:"frozen_duration"`
-	ChangeThreshold          int64  `json:"change_threshold"`
+	ChangeThreshold          float32  `json:"change_threshold"`
 }
 
 var Config ConfigObj
